Relocate comm code import source to importer app.
