+++ commodities/jinja2/commodities/import-success.jinja
@@ -9,8 +9,8 @@
   {{ govukBreadcrumbs({
     "items": [
       {"text": "Home", "href": url("home")},
-      {"text": "EU Taric Import List", "href": url("eu_import_batch-ui-list")},
-      {"text": "Import commodities", "href": url("commodity-ui-import")},
+      {"text": "EU Taric Import List", "href": url("commodity_importer-ui-list")},
+      {"text": "Import commodities", "href": url("commodity_importer-ui-create")},
       {"text": page_title}
     ]
   }) }}
