+++ commodities/jinja2/commodities/import.jinja
@@ -7,7 +7,7 @@
   {{ govukBreadcrumbs({
     "items": [
        {"text": "Home", "href": url("home")},
-    {"text": "EU Taric Import List", "href": url("eu_import_batch-ui-list")},
+    {"text": "EU Taric Import List", "href": url("commodity_importer-ui-list")},
     {"text": page_title}
     ]
   }) }}
