+++ commodities/tests/test_views.py
@@ -28,16 +28,16 @@ TEST_FILES_PATH = path.join(path.dirname(__file__), "test_files")
 
 
 def test_commodities_import_200(valid_user_client):
-    url = reverse("commodity-ui-import")
+    url = reverse("commodity_importer-ui-create")
     response = valid_user_client.get(url)
     assert response.status_code == 200
 
 
-@patch("commodities.forms.CommodityImportForm.save")
+@patch("importer.forms.CommodityImportForm.save")
 def test_commodities_import_success_redirect(mock_save, valid_user_client):
     mock_save.return_value = ImportBatchFactory.create()
-    url = reverse("commodity-ui-import")
-    redirect_url = reverse("commodity-ui-import-success")
+    url = reverse("commodity_importer-ui-create")
+    redirect_url = reverse("commodity_importer-ui-success")
     with open(f"{TEST_FILES_PATH}/valid.xml", "rb") as f:
         content = f.read()
     taric_file = SimpleUploadedFile("taric_file.xml", content, content_type="text/xml")
@@ -59,7 +59,7 @@ def test_commodities_import_success_redirect(mock_save, valid_user_client):
     ],
 )
 def test_commodities_import_failure(file_name, error_msg, valid_user_client):
-    url = reverse("commodity-ui-import")
+    url = reverse("commodity_importer-ui-create")
     with open(f"{TEST_FILES_PATH}/{file_name}", "rb") as f:
         content = f.read()
     taric_file = SimpleUploadedFile("taric_file.xml", content, content_type="text/xml")
