+++ common/jinja2/common/dashboard.jinja
@@ -85,7 +85,7 @@
       <h2 class="govuk-heading-m">Commodities</h2>
       <ul class="govuk-list">
         <li>
-          <a class="govuk-link" href="{{ url('commodity-ui-import') }}">
+          <a class="govuk-link" href="{{ url('commodity_importer-ui-create') }}">
             Import commodities
           </a>
         </li>
