+++ importer/jinja2/eu-importer/select-imports.jinja
@@ -23,12 +23,12 @@
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-one-third govuk-!-margin-bottom-4" style="float: right;">
       <a 
-      href="{{ url('commodity-ui-import') }}"
-      class="govuk-button "
-      style="float: right;"
-      data-module="govuk-button"
+        href="{{ url('commodity_importer-ui-create') }}"
+        class="govuk-button "
+        style="float: right;"
+        data-module="govuk-button"
       >
-      Import new Taric file
+        Import new Taric file
       </a>
     </div>
   </div>
