+++ importer/jinja2/includes/taric_importer_list.jinja
@@ -4,12 +4,17 @@
   {%- set object_status -%}
     <span class="status-badge">{{object.status}}</a>
   {%- endset -%}
+
+  {%- set object_link -%}
+    <a class="govuk-link" href="#">Resume</a>
+  {%- endset -%}
+
   {{ table_rows.append([
     {"text": object.name},
     {"text": "{:%d %b %Y}".format(object.created_at)},
     {"text": object.author.get_full_name() if object.author else "-"},
     {"text": object_status},
-    {"text": "Placeholder Resume Link"},
+    {"html": object_link},
   ]) or "" }}
 {% endfor %}
