+++ importer/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 
 from importer import views
 
-urlpatterns = [
+general_importer_urlpatterns = [
     path(
         "importer/",
         views.ImportBatchList.as_view(),
@@ -13,10 +13,32 @@ urlpatterns = [
         views.UploadTaricFileView.as_view(),
         name="import_batch-ui-create",
     ),
-    # new comm code importer url
+]
+
+commodity_importer_urlpatterns = [
     path(
-        "eu-importer/",
+        "commodity-importer/",
         views.TaricImportList.as_view(),
-        name="eu_import_batch-ui-list",
+        name="commodity_importer-ui-list",
+    ),
+    path(
+        "commodity-importer/create/",
+        # TODO
+        # Rename
+        views.CommodityImportView.as_view(),
+        # views.TaricImportCreateView.as_view(),
+        name="commodity_importer-ui-create",
+        # was commodity-ui-import
+    ),
+    path(
+        "commodity-importer/success/",
+        # TODO
+        # Rename
+        views.CommodityImportSuccessView.as_view(),
+        # views.TaricImportSuccessView.as_view(),
+        name="commodity_importer-ui-success",
+        # was commodity-ui-import-success
     ),
 ]
+
+urlpatterns = general_importer_urlpatterns + commodity_importer_urlpatterns
